AG-11148 Move memento into api/state/
